chore: rename Package to ls-mcmc
