+++ src/ls_mcmc/__init__.py
@@ -1,0 +1,1 @@
+"""Large Scale MCMC Library."""
