+++ src/ls_mcmc/algorithms.py
