+++ src/ls_mcmc/logging.py
@@ -1,4 +1,5 @@
 """Simple logger for tracking MCMC sampling."""
+
 import logging
 import sys
 from dataclasses import dataclass
