+++ src/ls_mcmc/model.py
