+++ src/ls_mcmc/output.py
